WIP / untested: remove ournode_{exit,end}_cmd
Recent versions of libosmocore register their own 'exit' and 'end'
commands to every node by default, so we *must* not register our own
version of those commands anymore.

This change is pending proper testing (do we really get into the right
nodes at 'exit' point as in the old commands?) as well as a dependency
on a higher libosmocore version tag!

Change-Id: I9d55b606ac3aa0c38814da569d07a7066ac89495