transceiver: Fix RSP SETSLOT format
According to osmo-trx README:
SETSLOT sets the format of the uplink timeslots in the ARFCN.
The <timeslot> indicates the timeslot of interest.
The <chantype> indicates the type of channel that occupies the timeslot.
A chantype of zero indicates the timeslot is off.
CMD SETSLOT <timeslot> <chantype>
RSP SETSLOT <status> <timeslot> <chantype>

Change-Id: I63c81155c84fc35cff1704734fd8d2732235c5a4
Related: OS#2944